Extract risk concentration mapper module
