+++ quality/refactor_health_report.md
@@ -7,8 +7,9 @@ Phase: baseline/report-only
 
 Recent gateway hardening has reduced monolithic Workbench, router-registry, performance workspace,
 advisor-brief, risk drawdown, risk rolling, and risk attribution responsibilities by extracting
-focused service adapters while preserving public behavior and keeping CI green. The remaining work
-is still substantial: large portfolio, risk workspace, contract, and client modules remain.
+focused service adapters and has extracted risk concentration response mapping behind a dedicated
+service module while preserving public behavior and keeping CI green. The remaining work is still
+substantial: large portfolio, risk workspace, contract, and client modules remain.
 
 ## Health Signals
 
@@ -28,8 +29,8 @@ is still substantial: large portfolio, risk workspace, contract, and client modu
 
 1. Split `portfolio_service.py` into source-readiness, transaction/activity, income, workspace,
    and workflow-cue adapters.
-2. Continue splitting `risk_workspace_service.py` by risk surface, with concentration and shared
-   unavailable-envelope helpers as the next risk workspace targets.
+2. Continue splitting `risk_workspace_service.py` by risk surface, with shared unavailable-envelope
+   helpers as the next risk workspace target.
 3. Split `platform_capabilities_service.py` capability normalization into smaller adapters.
 4. Continue extracting performance workspace evidence and attribution helpers behind stable
    response contracts.
